refactor: AlertManager 타입 구현
* Alertable 타입은 splitViewController를 참조해야하기 때문에 splitViewController를 weak var 로 참조하도록 구현
* AlertManager를 구현해 splitView의 역할 분리
* SplitViewController가 AlertManager를 지니고 이를 NoteListViewController에게 전달
