+++ CloudNotes/CloudNotes/Controller/SplitViewController.swift
@@ -7,9 +7,9 @@
 import UIKit
 
 class SplitViewController: UISplitViewController {
-    // MARK: Popover Properties
     var activityViewPopover: UIPopoverPresentationController?
     var actionSheetPopover: UIPopoverPresentationController?
+    lazy var alertManager = AlertManager(view: self)
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -29,6 +29,8 @@ class SplitViewController: UISplitViewController {
         
         setViewController(primaryChild, for: .primary)
         setViewController(secondaryChild, for: .secondary)
+        
+        primaryViewController.alertDelegate = alertManager
     }
     
     private func setUpSplitView() {
@@ -58,13 +60,11 @@ extension SplitViewController {
         case transition
     }
     
-    override func viewWillTransition(to size: CGSize,
-                                     with coordinator: UIViewControllerTransitionCoordinator)
+    override func viewWillTransition(to size: CGSize, with coordinator: UIViewControllerTransitionCoordinator)
     {
         super.viewWillTransition(to: size, with: coordinator)
 
-        configurePadSetting(for: actionSheetPopover, status: .transition)
-        configurePadSetting(for: activityViewPopover, status: .transition)
+        alertManager.configurePadTransition()
     }
     
     private func configurePadSetting(for popover: UIPopoverPresentationController?, status: ViewStatus) {
@@ -76,60 +76,5 @@ extension SplitViewController {
             popover?.sourceRect = newRect
             popover?.permittedArrowDirections = []
         }
-    }
-}
-
-// MARK: Alertable
-extension SplitViewController: Alertable {
-    func showActionSheet(of indexPath: IndexPath, noteTitle: String, deleteHandler: @escaping () -> Void) {
-        let actionSheet = UIAlertController(title: ActionSheet.title,
-                                            message: nil,
-                                            preferredStyle: .actionSheet)
-        
-        let shareAction = UIAlertAction(title: ActionSheet.shareAction, style: .default) { _ in
-            self.showActivityView(of: indexPath, noteTitle: noteTitle)
-        }
-        
-        let deleteAction = UIAlertAction(title: ActionSheet.deleteAction, style: .destructive) { _ in
-            self.showDeleteAlert(of: indexPath, deleteHandler: deleteHandler)
-        }
-        
-        let cancelAction = UIAlertAction(title: ActionSheet.cancelAction, style: .cancel)
-        
-        actionSheet.addAction(shareAction)
-        actionSheet.addAction(deleteAction)
-        actionSheet.addAction(cancelAction)
-        
-        configurePadSetting(for: actionSheet.popoverPresentationController, status: .load)
-        actionSheetPopover = actionSheet.popoverPresentationController
-
-        self.present(actionSheet, animated: true, completion: nil)
-    }
-    
-    func showActivityView(of indexPath: IndexPath, noteTitle: String) {
-        let shareText: String = noteTitle
-        let activityView = UIActivityViewController(activityItems: [shareText], applicationActivities: nil)
-        
-        configurePadSetting(for: activityView.popoverPresentationController, status: .load)
-        activityViewPopover = activityView.popoverPresentationController
-        
-        self.present(activityView, animated: true, completion: nil)
-    }
-    
-    func showDeleteAlert(of indexPath: IndexPath, deleteHandler: @escaping () -> Void) {
-        let alert = UIAlertController(title: Alert.title,
-                                      message: Alert.message,
-                                      preferredStyle: .alert)
-  
-        let cancelAction = UIAlertAction(title: Alert.cancelAction, style: .cancel)
-        
-        let deleteAction = UIAlertAction(title: Alert.deleteAction, style: .destructive) { _ in
-            deleteHandler()
-        }
-
-        alert.addAction(deleteAction)
-        alert.addAction(cancelAction)
-        
-        self.present(alert, animated: true, completion: nil)
     }
 }
