+++ CloudNotes/CloudNotes/Protocol/Alertable.swift
@@ -5,10 +5,13 @@
 //  Created by 홍정아 on 2021/09/15.
 //
 
-import Foundation
+import UIKit
 
 protocol Alertable: AnyObject {
+    var activityViewPopover: UIPopoverPresentationController? { get set }
+    var actionSheetPopover: UIPopoverPresentationController? { get set }
     func showActionSheet(of indexPath: IndexPath, noteTitle: String, deleteHandler: @escaping () -> Void)
     func showActivityView(of indexPath: IndexPath, noteTitle: String)
     func showDeleteAlert(of indexPath: IndexPath, deleteHandler: @escaping () -> Void)
+    func configurePadTransition()
 }
